Add files via upload
Simple BLDC Motor Controller Circuit Using Irfz44n Mosfet

 

IRFZ44N Mosfet

IRFZ44 is one of the N Channel Mosfet Types and is in TO-220 sheath. IRFZ44 is N Channel Power Mosfet type. The IRFZ44 has a Gate-Source voltage of 60V and has a continuous output current of 48A.

N type channel in this MOSFET, As the name suggests. In addition, it has other technical details such as:

-Drain source disconnection voltage: 60 V

-Continuous discharge intensity: 50A

-RDS: 22mOhm

-Gate source voltage: 20 V

-Operating temperature range: -55 to 175ºC

-Power dissipation: 131w

-Falling time: 13 ns

-Establishment time: 55 ns

-Power-off delay: 37 ns

-Typical link delay: 12ns





PART LIST___



1) 3x IRFZ44N / 3205 Mosfets

2) 6x 1N4007 Diodes

3) 3x 10 Kilo Ohms Resistors

4) 1 Kilo Ohms Resistor

5) 100k potentiometer